+++ 3rdparty/rudeconfig/ParserJuly2004.cpp
@@ -67,11 +67,11 @@ enum KeyValueState{ KEY, KEYESCAPE, STARTVALUE, COMMENT, FINDCOMMENT, KVERROR, E
 
 void ParserJuly2004::stripTrailing(std::string& buffer)
 {
-	int bufferLength = buffer.size();
+	size_t bufferLength = buffer.size();
 
 	locale loc;
 
-	for (int x = bufferLength - 1; x >= 0; x--)
+	for (size_t x = bufferLength - 1; x >= 0; x--)
 	{
 		char c = buffer[x];
